pyproject.toml
[project]
name = "chomikai"
version = "0.1.0"
description = "MCP tool which connects to Google Slides and generates a presentation."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "fastapi>=0.115.12",
    "google-api-python-client>=2.166.0",
    "google-api-python-client-stubs>=1.29.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "mcp[cli]>=1.6.0",
    "pydantic>=2.11.2",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.15.0",
    "poethepoet>=0.33.1",
    "pylint>=3.3.6",
    "pyright>=1.1.398",
    "ruff>=0.11.4",
    "types-cffi>=1.17.0.20250326",
    "types-pyopenssl>=24.1.0.20240722",
    "types-simplejson>=3.20.0.20250326",
    "types-ujson>=5.10.0.20250326",
]
